Issue 15 - redraw review score view as 2 halves so we can render float values.
Update scores to be floats and update the test values in PlaceCarousel to test half score rendering

+++ Prox/Prox/Widgets/ReviewContainerView.swift
@@ -13,7 +13,7 @@ class ReviewContainerView: UIView {
         }
     }
 
-    var score: Int = 0 {
+    var score: Float = 0 {
         didSet {
             reviewScore.score = score
         }
@@ -43,7 +43,7 @@ class ReviewContainerView: UIView {
         self.init(score: 0)
     }
 
-    init(score: Int) {
+    init(score: Float) {
         super.init(frame: .zero)
         self.score = score
         setupSubviews()
